Remove StoreUtils
- move the operations inside the stat suppliers.
